requirements.txt
# Banking RAG System Requirements

# Core AI and ML libraries
openai>=1.0.0
numpy>=1.21.0
faiss-cpu>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Web server dependencies
flask>=2.3.0
flask-cors>=4.0.0

# Environment and configuration
python-dotenv>=1.0.0

# For GPU support (optional):
# faiss-gpu>=1.7.0

# Production WSGI server
gunicorn>=21.0.0

# Additional utilities
pickle-mixin>=1.0.2
typing-extensions>=4.0.0
requests>=2.31.0
